+++ tests/test_model.py
@@ -11,16 +11,16 @@ from solvis_store import model
 class PynamoTest(unittest.TestCase):
     def setUp(self):
         model.set_local_mode()
-        model.RuptureSetLocationRadiusRuptures.create_table(wait=True)
-        print("Migrate created table: RuptureSetLocationRadiusRuptures")
+        model.RuptureSetLocationDistances.create_table(wait=True)
+        print("Migrate created table: RuptureSetLocationDistances")
 
     def test_table_exists(self):
         # with app.app_context():
-        self.assertEqual(model.RuptureSetLocationRadiusRuptures.exists(), True)
+        self.assertEqual(model.RuptureSetLocationDistances.exists(), True)
 
     def tearDown(self):
         # with app.app_context():
-        model.RuptureSetLocationRadiusRuptures.delete_table()
+        model.RuptureSetLocationDistances.delete_table()
         return super(PynamoTest, self).tearDown()